Generate the for_each_hegel_fn! list from hegel-c
The frontend's list of hegel_* functions and their signatures was
maintained by hand in src/ffi/sys.rs. The static-engine drift test could
only check what was listed, so a function added to the engine and called
from src/ffi.rs but missing from the list compiled fine under
--all-features and failed only in a default-features build, as happened
on #476.

scripts/gen-ffi-list.py now extracts every no_mangle pub extern "C" fn
hegel_* signature from hegel-c/src/lib.rs and writes the macro to
src/ffi/sys/fns.rs, the same way cbindgen derives hegel.h. `just
c-header` regenerates both, and the new check-ffi-list recipe in
check-lint fails when the file is stale. The list is now the whole
exported ABI, so hegel_version and the four functions the frontend does
not yet call get loader wrappers too, which are allowed to be dead; the
drift test's separate hegel_version assertion is folded into the
generated list.
