+++ AGENTS.md
@@ -28,7 +28,7 @@ MSRV is 1.86 (enforced in CI and Cargo.toml). If you bump it, also bump `ci.yml`
 
 - `src/lib.rs` — Public API surface: `hegel()`, the `Hegel` builder, `TestCase`, the `Generator` trait, and the proc-macro re-exports (`#[hegel::test]`, `#[hegel::main]`, `#[derive(DefaultGenerator)]`, `#[composite]`, `#[state_machine]`, `#[reproduce_failure]`, …)
 - `src/ffi.rs` — The libhegel C-ABI boundary: the only module that touches the raw `hegel_*` functions; the rest of the frontend works against its safe wrappers (`SettingsHandle`, `RunHandle`, `CTestCase`, `RunResult`)
-- `src/ffi/sys.rs` — Where the `hegel_*` functions come from, in two interchangeable forms: by default mirrored `repr(C)` types (`sys/types.rs`) plus a runtime loader (`sys/loader.rs`) that dlopens the `libhegel_c` cdylib `build.rs` built, so `hegeltest-c` and its dependencies stay out of consumers' cargo graphs; with the public `static-engine` feature, a re-export of the `hegel_c` rlib (self-contained binaries, visible engine dependencies — also what Miri and the coverage workspace pass use). One macro, `for_each_hegel_fn!`, is the single list of functions; embedded drift tests compile-assert it and the mirrors against `hegel_c` under `static-engine`
+- `src/ffi/sys.rs` — Where the `hegel_*` functions come from, in two interchangeable forms: by default mirrored `repr(C)` types (`sys/types.rs`) plus a runtime loader (`sys/loader.rs`) that dlopens the `libhegel_c` cdylib `build.rs` built, so `hegeltest-c` and its dependencies stay out of consumers' cargo graphs; with the public `static-engine` feature, a re-export of the `hegel_c` rlib (self-contained binaries, visible engine dependencies — also what Miri and the coverage workspace pass use). One macro, `for_each_hegel_fn!` in `sys/fns.rs`, is the single list of functions, generated from `hegel-c/src/lib.rs` by `scripts/gen-ffi-list.py` (`just c-header` refreshes it, `just check-ffi-list` checks it); embedded drift tests compile-assert it and the mirrors against `hegel_c` under `static-engine`
 - `src/run_lifecycle.rs` — Cross-cutting per-test-case lifecycle: panic hook, `catch_unwind` wrapping, translating panics into `TestCaseResult`, and the final re-raise
 - `src/backend.rs` — The result types the lifecycle speaks (`TestCaseResult`, `Failure`)
 - `src/test_case.rs` — `TestCase` (the handle test bodies draw from) and its thread-local state, the `Collection` helper, and the span `labels` module
