+++ RELEASE.md
@@ -1,0 +1,3 @@
+RELEASE_TYPE: patch
+
+Internal refactoring of how the frontend declares the engine's C ABI.
